requirements.txt
configparser==3.5.0
elasticsearch==5.4.0
python-dateutil==2.6.1
python-dotenv==0.7.1
tweepy==3.5.0